# FMS/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Force Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whitelist:
    get:
      operationId: listWhitelist
      tags: [whitelist]
      summary: List whitelist entries
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of whitelist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhitelistEntry"
    post:
      operationId: createWhitelistEntry
      tags: [whitelist]
      summary: Add a whitelist entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhitelistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhitelistEntry"
  /whitelist/{id}:
    patch:
      operationId: updateWhitelistEntry
      tags: [whitelist]
      summary: Update a whitelist entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhitelistUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhitelistEntry"
    delete:
      operationId: deleteWhitelistEntry
      tags: [whitelist]
      summary: Remove from whitelist
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # Careers
...
components:
  schemas:
...
    WhitelistEntry:
      type: object
      required: [id, discordId, status]
      properties:
        id: { type: integer }
        discordId: { type: string }
        discordUsername: { type: ["string", "null"] }
        characterName: { type: ["string", "null"] }
        status: { type: string }
        addedBy: { type: ["string", "null"] }
        addedAt: { type: string }
        notes: { type: ["string", "null"] }
    WhitelistInput:
      type: object
      required: [discordId, status]
      properties:
        discordId: { type: string }
        discordUsername: { type: string }
        characterName: { type: string }
        status: { type: string }
        addedBy: { type: string }
        notes: { type: string }
    WhitelistUpdate:
      type: object
      properties:
        discordUsername: { type: string }
        characterName: { type: string }
        status: { type: string }
        addedBy: { type: string }
        notes: { type: string }

    # Career schemas
...