# FMS/lib/api-zod/src/generated/types/whitelistUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Force Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface WhitelistUpdate {
  discordUsername?: string;
  characterName?: string;
  status?: string;
  addedBy?: string;
  notes?: string;
}